Allow users to open the filesystem with mode as 'wb' or 'rb'
Currently users can only open the filesystem using the mode 'r' or
'w'. Since the file system is in binary mode (and not text mode) users
will expect that they can open using 'rb' or 'wb'.

This adds the ability to use 'rb' or 'wb' as the mode. Which act the
same as 'r' or 'w' respectively.